Moving endpoint files
