+++ apps/meteor/app/livechat/server/api.ts
@@ -1,5 +1,4 @@
 import '../imports/server/rest/agent';
-import '../imports/server/rest/departments';
 import '../imports/server/rest/facebook';
 import '../imports/server/rest/sms.js';
 import '../imports/server/rest/users';
