+++ apps/meteor/ee/app/livechat-enterprise/server/api/commonApiIndex.ts
@@ -1,0 +1,3 @@
+// Although these api's are defined in the enterprise package, they are still available in the community edition.
+
+import './departments';
